requirements.txt
ibm-watsonx-ai>=1.3.32
ibm-watson>=10.0.0
pysqlite3-binary
chromadb>=1.0.15
pypdf>=4.0.0
python-docx>=1.2.0
openpyxl>=3.1.0
pandas>=2.2.0
pytesseract>=0.3.13
Pillow>=10.0.0
pdf2image<=1.17.0
reportlab>=4.0.0
fastapi>=0.104.0
uvicorn>=0.23.0